+++ lib/features/og_meta_data/og_image.g.dart
@@ -8,7 +8,7 @@ part of 'og_image.dart';
 
 OGImage _$OGImageFromJson(Map<String, dynamic> json) {
   return OGImage(
-    url: json['url'] as String,
+    url: json['url'] as String?,
     secureUrl: json['secure_url'] as String?,
     type: json['type'] as String?,
     alt: json['alt'] as String?,