+++ lib/features/og_meta_data/og_interfaces.dart
@@ -1,5 +1,5 @@
 abstract class OGMedia {
-  final String url = '';
+  final String? url = '';
   final String? secureUrl = '';
 }